# docs/backend/backend_python/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PictoPy",
    "description": "The API calls to PictoPy are done via HTTP requests. This backend is built using FastAPI.",
    "version": "0.1.0",
    "contact": {
      "name": "PictoPy Postman Collection",
      "url": "https://www.postman.com/aossie-pictopy/pictopy/overview"
    }
  },
  "servers": [
    {
      "url": "http://localhost:52123",
      "description": "Local Development server"
    }
  ],
  "paths": {
...
    "/albums/{album_id}/images": {
      "post": {
        "tags": [
          "Albums"
        ],
        "summary": "Add Images To Album",
        "operationId": "add_images_to_album_albums__album_id__images_post",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Album Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageIdsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Albums"
        ],
        "summary": "Remove Images From Album",
        "operationId": "remove_images_from_album_albums__album_id__images_delete",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Album Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageIdsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/albums/{album_id}/images/{image_id}": {
      "delete": {
        "tags": [
          "Albums"
        ],
        "summary": "Remove Image From Album",
        "operationId": "remove_image_from_album_albums__album_id__images__image_id__delete",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Album Id"
            }
          },
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Get All Images",
        "description": "Get all images from the database.",
        "operationId": "get_all_images_images__get",
        "parameters": [
          {
            "name": "tagged",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter images by tagged status",
              "title": "Tagged"
            },
            "description": "Filter images by tagged status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllImagesResponse"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__images__ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/toggle-favourite": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Toggle Favourite",
        "operationId": "toggle_favourite_images_toggle_favourite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleFavouriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/download/{image_id}": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Download Image",
        "description": "Download an image file.",
        "operationId": "download_image_images_download__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/face-clusters/{cluster_id}/images": {
      "get": {
        "tags": [
          "Face Clusters"
        ],
        "summary": "Get Cluster Images",
        "description": "Get all images that contain faces belonging to a specific cluster.",
        "operationId": "get_cluster_images_face_clusters__cluster_id__images_get",
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetClusterImagesResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__face_clusters__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__face_clusters__ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetAllImagesResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ImageData"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "title": "GetAllImagesResponse"
      },
      "GetClusterImagesData": {
        "properties": {
          "cluster_id": {
            "type": "string",
            "title": "Cluster Id"
          },
          "cluster_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Name"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageInCluster"
            },
            "type": "array",
            "title": "Images"
          },
          "total_images": {
            "type": "integer",
            "title": "Total Images"
          }
        },
        "type": "object",
        "required": [
          "cluster_id",
          "images",
          "total_images"
        ],
        "title": "GetClusterImagesData",
        "description": "Data model for cluster images response."
      },
      "GetClusterImagesResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GetClusterImagesData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "GetClusterImagesResponse",
        "description": "Response model for getting images in a cluster."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "folder_id": {
            "type": "string",
            "title": "Folder Id"
          },
          "thumbnailPath": {
            "type": "string",
            "title": "Thumbnailpath"
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataModel"
          },
          "isTagged": {
            "type": "boolean",
            "title": "Istagged"
          },
          "isFavourite": {
            "type": "boolean",
            "title": "Isfavourite"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "path",
          "folder_id",
          "thumbnailPath",
          "metadata",
          "isTagged",
          "isFavourite"
        ],
        "title": "ImageData"
      },
      "ImageIdsRequest": {
        "properties": {
          "image_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Image Ids"
          }
        },
        "type": "object",
        "required": [
          "image_ids"
        ],
        "title": "ImageIdsRequest"
      },
      "ImageInCluster": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "thumbnailPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailpath"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "face_id": {
            "type": "integer",
            "title": "Face Id"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "bbox": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          }
        },
        "type": "object",
        "required": [
          "id",
          "path",
          "face_id"
        ],
        "title": "ImageInCluster",
        "description": "Represents an image that contains faces from a specific cluster."
      },
...
      "MetadataModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "date_created": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Created"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "file_location": {
            "type": "string",
            "title": "File Location"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "item_type": {
            "type": "string",
            "title": "Item Type"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "name",
          "date_created",
          "width",
          "height",
          "file_location",
          "file_size",
          "item_type"
        ],
        "title": "MetadataModel"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          }
        },
        "type": "object",
        "required": [
          "success",
          "msg"
        ],
        "title": "SuccessResponse"
      },
...
      "ToggleFavouriteRequest": {
        "properties": {
          "image_id": {
            "type": "string",
            "title": "Image Id"
          }
        },
        "type": "object",
        "required": [
          "image_id"
        ],
        "title": "ToggleFavouriteRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "app__schemas__face_clusters__ErrorResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "ErrorResponse"
      },
...
      "app__schemas__images__ErrorResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "error": {
            "type": "string",
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "message",
          "error"
        ],
        "title": "ErrorResponse"
      },
...
    }
  }
}